refactor: BatchNames 패키지 이동
