+++ backend/application-modules/batch-server/src/main/java/com/softeer/batch/forecast/shortterm/config/ShortForecastJobConfig.java
@@ -20,9 +20,9 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.transaction.PlatformTransactionManager;
 
-import static com.softeer.batch.forecast.support.BatchNames.Handlers.SHORT_SIMPLE_RETRY_HANDLER;
-import static com.softeer.batch.forecast.support.BatchNames.Steps.SCHEDULED_SHORT_FORECAST_STEP;
-import static com.softeer.batch.forecast.support.BatchNames.Steps.START_UP_SHORT_FORECAST_STEP;
+import static com.softeer.batch.common.support.BatchNames.Handlers.SHORT_SIMPLE_RETRY_HANDLER;
+import static com.softeer.batch.common.support.BatchNames.Steps.SCHEDULED_SHORT_FORECAST_STEP;
+import static com.softeer.batch.common.support.BatchNames.Steps.START_UP_SHORT_FORECAST_STEP;
 
 @Slf4j
 @Configuration
